+++ build.rs
@@ -1,10 +1,5 @@
-use std::path::PathBuf;
-use std::process::Command;
-
-#[allow(dead_code)]
-const MIN_LIBHEIF_VERSION: &str = "1.19";
-#[allow(dead_code)]
-const LIBHEIF_GITHUB_VERSION: &str = "v1.19.7";
+use std::io;
+use std::path::{Path, PathBuf};
 
 fn main() {
     if std::env::var("DOCS_RS").is_ok() {
@@ -13,17 +8,12 @@ fn main() {
     }
 
     println!("cargo:rerun-if-changed=build.rs");
-    println!("cargo:rerun-if-changed=include/wrapper.h");
+    println!("cargo:rerun-if-changed=wrapper.h");
 
     // Tell cargo to tell rustc to link the heif library.
 
     #[cfg(not(target_os = "windows"))]
-    {
-        #[cfg(feature = "compile-libheif")]
-        compile_libheif();
-
-        find_libheif();
-    }
+    find_libheif();
 
     #[cfg(target_os = "windows")]
     install_libheif_by_vcpkg();
@@ -33,76 +23,29 @@ fn main() {
 }
 
 #[allow(dead_code)]
-fn run_command(description: &str, cmd: &mut Command) -> String {
-    match cmd.output() {
-        Ok(output) if !output.status.success() => {
-            let std_err = String::from_utf8_lossy(&output.stderr);
-            println!(
-                "cargo:warning=Failed {}.\n{}\n{}",
-                description, output.status, std_err
-            );
-            std::process::exit(1);
-        }
-        Err(e) => {
-            println!("cargo:warning={}", e);
-            std::process::exit(1);
-        }
-        Ok(output) => String::from_utf8_lossy(&output.stdout).to_string(),
-    }
-}
-
-#[allow(dead_code)]
-fn fetch_libheif() -> PathBuf {
+fn prepare_libheif_src() -> PathBuf {
     let out_path = PathBuf::from(std::env::var("OUT_DIR").unwrap());
-    let libheif_dir = out_path.join("libheif");
-    if libheif_dir.exists() {
-        let stdout = run_command(
-            "getting 'libheif' version from exists git repo",
-            Command::new("git").args([
-                "-C",
-                libheif_dir.to_str().unwrap(),
-                "describe",
-                "--exact-match",
-                "--tags",
-            ]),
-        );
-        if stdout.trim() != LIBHEIF_GITHUB_VERSION {
-            std::fs::remove_dir_all(&libheif_dir).unwrap();
-        }
-    }
-
-    if !libheif_dir.exists() {
-        run_command(
-            "fetching 'libheif' from GitHub",
-            Command::new("git").args([
-                "clone",
-                "--depth",
-                "1",
-                "--branch",
-                LIBHEIF_GITHUB_VERSION,
-                "https://github.com/strukturag/libheif.git",
-                libheif_dir.to_str().unwrap(),
-            ]),
-        );
-    }
-
-    libheif_dir
-}
-
-#[cfg(feature = "compile-libheif")]
-fn compile_libheif() {
-    use std::path::PathBuf;
-
-    let out_path = PathBuf::from(std::env::var("OUT_DIR").unwrap());
-    let libheif_dir = fetch_libheif();
+    let crate_dir = PathBuf::from(std::env::var("CARGO_MANIFEST_DIR").unwrap());
+    let libheif_dir = crate_dir.join("vendor/libheif");
+    let dst_dir = out_path.join("libheif");
+    copy_dir_all(libheif_dir, &dst_dir).unwrap();
 
     // Patch CMakeLists.txt to disable a building `heifio` library
     // that is used for example applications.
-    let cmake_lists_path = libheif_dir.join("CMakeLists.txt");
+    let cmake_lists_path = dst_dir.join("CMakeLists.txt");
     let mut contents =
         std::fs::read_to_string(&cmake_lists_path).expect("failed to read libheif/CMakeLists.txt");
     contents = contents.replace("add_subdirectory(heifio)", "");
     std::fs::write(&cmake_lists_path, contents).expect("failed to write libheif/CMakeLists.txt");
+    dst_dir
+}
+
+#[cfg(feature = "embedded-libheif")]
+fn compile_libheif() -> String {
+    use std::path::PathBuf;
+
+    let out_path = PathBuf::from(std::env::var("OUT_DIR").unwrap());
+    let libheif_dir = prepare_libheif_src();
 
     let mut build_config = cmake::Config::new(libheif_dir);
     build_config.out_dir(out_path.join("libheif_build"));
@@ -114,14 +57,13 @@ fn compile_libheif() {
         "WITH_GDK_PIXBUF",
         "WITH_EXAMPLES",
         "ENABLE_EXPERIMENTAL_FEATURES",
+        "ENABLE_PLUGIN_LOADING",
     ] {
         build_config.define(key, "OFF");
     }
 
     // Enable some options
     for key in [
-        "ENABLE_PLUGIN_LOADING",
-        "WITH_OPENJPH_ENCODER",
         "WITH_REDUCED_VISIBILITY",
         "WITH_LIBSHARPYUV",
         // TODO: Try to enable this in the future.
@@ -145,6 +87,7 @@ fn compile_libheif() {
         "JPEG_DECODER",
         "JPEG_ENCODER",
         "KVAZAAR",
+        "OPENJPH_ENCODER",
         "OpenJPEG_DECODER",
         "OpenJPEG_ENCODER",
         "FFMPEG_DECODER",
@@ -154,40 +97,62 @@ fn compile_libheif() {
         "VVENC",
     ];
 
-    // Enable encoders and decoders
-    for key in encoders_decoders {
-        build_config.define(format!("WITH_{}", key), "ON");
-    }
+    let disabled_enc_dec = [
+        // A lot of errors like "undefined reference to `av_packet_free'" during linking
+        "FFMPEG_DECODER",
+    ];
 
-    // Disable external plugins
-    #[cfg(feature = "embedded-libheif-plugins")]
-    const PLUGIN_OPTION_VALUE: &str = "OFF";
-    // Enable external plugins
-    #[cfg(not(feature = "embedded-libheif-plugins"))]
-    const PLUGIN_OPTION_VALUE: &str = "ON";
-
+    // Enable or disable encoders and decoders
     for key in encoders_decoders {
-        build_config.define(format!("WITH_{}_PLUGIN", key), PLUGIN_OPTION_VALUE);
+        let v = if disabled_enc_dec.contains(&key) {
+            "OFF"
+        } else {
+            "ON"
+        };
+        build_config.define(format!("WITH_{}", key), v);
+
+        // Disable external plugin
+        build_config.define(format!("WITH_{}_PLUGIN", key), "OFF");
     }
 
     let libheif_build = build_config.build();
 
-    let pkgconfig_dir = libheif_build
+    libheif_build
         .join("lib/pkgconfig")
         .to_string_lossy()
-        .to_string();
-    std::env::set_var("PKG_CONFIG_PATH", pkgconfig_dir);
+        .to_string()
 }
 
 #[cfg(not(target_os = "windows"))]
 fn find_libheif() {
-    let mut config = pkg_config::Config::new();
-    config.atleast_version(MIN_LIBHEIF_VERSION);
-    #[cfg(feature = "compile-libheif")]
-    config.statik(true);
+    #[allow(unused_mut)]
+    let mut config = system_deps::Config::new();
 
-    if let Err(err) = config.probe("libheif") {
-        println!("cargo:warning={}", err);
+    #[cfg(feature = "embedded-libheif")]
+    {
+        std::env::set_var("SYSTEM_DEPS_LIBHEIF_BUILD_INTERNAL", "always");
+        config = config.add_build_internal("libheif", |lib, version| {
+            let pc_file_path = compile_libheif();
+            system_deps::Library::from_internal_pkg_config(pc_file_path, lib, version)
+        });
+    }
+
+    use system_deps::Error;
+
+    if let Err(err) = config.probe() {
+        let err_msg = match &err {
+            Error::InvalidMetadata(msg) => {
+                if msg.contains("No version") && msg.contains("libheif") {
+                    "You MUST enable one of the crate features to specify \
+                    minimal supported version of 'libheif' API (e.g. v1_17)."
+                        .to_string()
+                } else {
+                    err.to_string()
+                }
+            }
+            _ => err.to_string(),
+        };
+        println!("cargo:error={}", err_msg);
         std::process::exit(1);
     }
 }
@@ -206,7 +171,7 @@ fn install_libheif_by_vcpkg() {
 #[cfg(feature = "use-bindgen")]
 fn run_bindgen() {
     let mut base_builder = bindgen::Builder::default()
-        .header("include/wrapper.h")
+        .header("wrapper.h")
         .generate_comments(true)
         .formatter(bindgen::Formatter::Rustfmt)
         .generate_cstr(true)
@@ -292,4 +257,18 @@ fn run_bindgen() {
     bindings
         .write_to_file(out_path.join("bindings_wo_tests.rs"))
         .expect("Couldn't write bindings_wo_tests.rs!");
+}
+
+fn copy_dir_all(src: impl AsRef<Path>, dst: impl AsRef<Path>) -> io::Result<()> {
+    std::fs::create_dir_all(&dst)?;
+    for entry in std::fs::read_dir(src)? {
+        let entry = entry?;
+        let ty = entry.file_type()?;
+        if ty.is_dir() {
+            copy_dir_all(entry.path(), dst.as_ref().join(entry.file_name()))?;
+        } else {
+            std::fs::copy(entry.path(), dst.as_ref().join(entry.file_name()))?;
+        }
+    }
+    Ok(())
 }