+++ include/wrapper.h
@@ -1,3 +1,0 @@
-#include "libheif/heif.h"
-#include "libheif/heif_properties.h"
-#include "libheif/heif_regions.h"