+++ wrapper.h
@@ -1,0 +1,3 @@
+#include "vendor/libheif/libheif/api/libheif/heif.h"
+#include "vendor/libheif/libheif/api/libheif/heif_properties.h"
+#include "vendor/libheif/libheif/api/libheif/heif_regions.h"